+++ src/functions/report/index.ts
@@ -1,5 +1,21 @@
 import { handlerPath } from "@libs/handler-resolver";
 
+//report generation
+export const generateReport = {
+    handler: `${handlerPath(__dirname)}/handler.generateReport`,
+    description: 'A function that generates a report.',
+    events: [
+        {
+            http: {
+                method: 'post',
+                path: 'generateReport/',
+                cors: true
+            },
+        },
+    ],
+};
+
+//retreival of all reports
 export const getAllReports = {
     handler: `${handlerPath(__dirname)}/handler.getAllReports`,
     description: 'A function that returns all of the reports in the system.',
@@ -14,6 +30,7 @@ export const getAllReports = {
     ],
 };
 
+//retrieval of drafts
 export const getAllMyReports = {
     handler: `${handlerPath(__dirname)}/handler.getAllMyReports`,
     description: 'A function that returns all the users reports.',
@@ -28,6 +45,7 @@ export const getAllMyReports = {
     ],
 };
 
+//retrevial of a report
 export const getReport = {
     handler: `${handlerPath(__dirname)}/handler.getReport`,
     description: 'A function that returns the content of a specific report.',
@@ -42,6 +60,7 @@ export const getReport = {
     ],
 };
 
+//cloning of a report
 export const cloneReport = {
     handler: `${handlerPath(__dirname)}/handler.cloneReport`,
     description: 'A function that clones a report.',
